0.0

+++ app/src/main/java/cjx/com/diary/util/WeightUtils.java
@@ -21,6 +21,7 @@ import cjx.com.diary.mode.weight.BodyWeightBeanDao;
 public class WeightUtils {
     /**
      * 获取所有数据列表
+     *
      * @return
      */
     public static List<BodyWeightBean> getWeightList() {
@@ -78,7 +79,7 @@ public class WeightUtils {
     public static String TYPE_NIGHT = "type_night";
 
     private static final String[] weeks = new String[]{
-            "周日", "周一", "周二", "周三", "周四", "周五","週六"
+            "周日", "周一", "周二", "周三", "周四", "周五", "週六"
     };
 
     public static final IAxisValueFormatter formatter = (value, axis) -> weeks[Integer.valueOf((int) value)];
@@ -106,7 +107,10 @@ public class WeightUtils {
         return rList;
     }
 
-    public static void clearData(){
-       MyApplication.INSTANCE.getDaoSession().getBodyWeightBeanDao().deleteAll();
+    public static void clearData() {
+        BodyWeightBeanDao dao = MyApplication.INSTANCE.getDaoSession().getBodyWeightBeanDao();
+        if (dao != null && dao.count() > 0) {
+            MyApplication.INSTANCE.getDaoSession().getBodyWeightBeanDao().deleteAll();
+        }
     }
 }
